+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/libdeprecated/CssDefaults.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2020 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.css;
+package com.diffplug.gradle.spotless.libdeprecated;
 
 import java.util.Arrays;
 import java.util.Collections;
